+++ src/views/components/world/map/MapMenu.tsx
@@ -12,7 +12,7 @@ import { createMapInfo } from '@utils/entityCreation';
 import { MAP_INFO_FOLDER_NAME_TEXT_ID, StudioMapInfoFolder } from '@modelEntities/mapInfo';
 import { useSetProjectText } from '@utils/ReadingProjectText';
 import { SeparatorGreyLine } from '@components/separators/SeparatorGreyLine';
-import { MapTreeV2 } from './tree/MapTreeV2';
+import { MapTree } from './tree/MapTree';
 import { addNewMapInfo } from '@utils/MapInfoUtils';
 
 const MapMenuContainer = styled(NavigationDatabaseStyle)`
@@ -63,7 +63,7 @@ export const MapMenu = () => {
             <NewFolderButtonOnlyIcon onClick={handleNewFolder} />
           </div>
           <SeparatorGreyLine />
-          <MapTreeV2 />
+          <MapTree />
         </MapSubMenuContainer>
       </NavigationDatabaseGroup>
       <MapEditorOverlay ref={dialogsRef} />
